+++ packages/veui/src/components/PromptBox.vue
@@ -7,8 +7,11 @@
   @ok="submit"
   escapable
   @escape="cancel"
-  @cancel="cancel">
-  <template slot="title"><slot name="title">{{ title }}</slot></template>
+  @cancel="cancel"
+  role="alertdialog">
+  <template slot="title">
+    <slot name="title">{{ title }}</slot>
+  </template>
   <p class="veui-prompt-box-info">{{ content }}</p>
   <veui-input autofocus v-model="localValue" class="veui-prompt-box-input" @keydown.enter="submit"></veui-input>
 </veui-dialog>